Before:

|    Method | LoopNum |       Mean |      Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------- |-------- |-----------:|-----------:|----------:|------------:|------------:|------------:|--------------------:|
| EmptyTask |     100 |   5.574 ms |  0.5928 ms | 0.1539 ms |    242.1875 |           - |           - |             7.98 KB |
|      Echo |     100 |  16.646 ms |  2.9067 ms | 0.7549 ms |    843.7500 |           - |           - |             8.17 KB |
| EmptyTask |    1000 |  52.223 ms |  5.0958 ms | 1.3234 ms |   2363.6364 |           - |           - |             7.98 KB |
|      Echo |    1000 | 158.258 ms | 10.9308 ms | 2.8387 ms |   8000.0000 |           - |           - |             8.17 KB |

After:

|    Method | LoopNum |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------- |-------- |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| EmptyTask |     100 |   5.559 ms | 0.9527 ms | 0.2474 ms |    242.1875 |           - |           - |             7.98 KB |
|      Echo |     100 |  12.640 ms | 1.1351 ms | 0.2948 ms |    781.2500 |           - |           - |             8.17 KB |
| EmptyTask |    1000 |  52.710 ms | 9.8790 ms | 2.5655 ms |   2400.0000 |           - |           - |             7.98 KB |
|      Echo |    1000 | 126.787 ms | 7.4302 ms | 1.9296 ms |   7750.0000 |           - |           - |             8.17 KB |